+++ src/test/converter2/validation/externalType.ts
@@ -1,0 +1,2 @@
+import { Program } from "typescript";
+export declare const program: Program;